% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readvcf.R
\name{getsnpinfo}
\alias{getsnpinfo}
\title{Read SNP Information Without Genotype Data}
\usage{
getsnpinfo(vcf_file, genome = "hg19", verbose = TRUE)
}
\arguments{
\item{vcf_file}{Character string with path to VCF file (must be bgzipped)}

\item{genome}{Character string specifying genome build (e.g., "hg19", "hg38")}

\item{verbose}{Logical; if TRUE, display informational messages. Default is TRUE.}
}
\value{
A VCF object containing SNP information (position, alleles, INFO fields) without genotype data
}
\description{
Read SNP Information Without Genotype Data
}
\examples{
\dontrun{
vcf_file <- "data/axiom_acs_aus_nf_chr21.vcf.gz"
snps <- getsnpinfo(vcf_file, genome = "hg19")
# Access SNP information
rowRanges(snps)  # Genomic positions and alleles
info(snps)       # INFO fields (AF, MAF, R2, etc.)
}
}
