% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readvcf.R
\name{getvcfheader}
\alias{getvcfheader}
\title{Read VCF Header Information}
\usage{
getvcfheader(vcf_file, check_tabix = TRUE, verbose = TRUE)
}
\arguments{
\item{vcf_file}{Character string with path to VCF file (must be bgzipped)}

\item{check_tabix}{Logical; if TRUE, checks for tabix index file}

\item{verbose}{Logical; if TRUE, display informational messages. Default is TRUE.}
}
\value{
A VCFHeader object containing header information
}
\description{
Read VCF Header Information
}
\examples{
\dontrun{
vcf_file <- "data/axiom_acs_aus_nf_chr21.vcf.gz"
header <- getvcfheader(vcf_file)
# Access header components
samples(header)
meta(header)
geno(header)
}
}
