% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readvcf.R
\name{getvcfsnp}
\alias{getvcfsnp}
\title{Read Dosage (DS) Genotype Information for SNPs}
\usage{
getvcfsnp(vcf_summary, snp_indices, verbose = FALSE)
}
\arguments{
\item{vcf_summary}{List returned from summarizevcf() function containing vcf_file, genome, snp_summary, and sample_ids}

\item{snp_indices}{Integer vector of SNP indices to retrieve (1-based). Can be a single index or multiple indices.}

\item{verbose}{Logical; if TRUE, display informational messages during execution. Default is FALSE.}
}
\value{
If single SNP requested, a numeric vector of DS values. If multiple SNPs, a matrix (rows = SNPs, columns = samples).
}
\description{
Read Dosage (DS) Genotype Information for SNPs
}
\examples{
\dontrun{
vcf_file <- "data/axiom_acs_aus_nf_chr21.vcf.gz"
vcf_summary <- summarizevcf(vcf_file)

# Read a single SNP
dosage <- getvcfsnp(vcf_summary, snp_indices = 1)

# Read multiple SNPs efficiently
dosages <- getvcfsnp(vcf_summary, snp_indices = c(1, 2, 5, 10))

# Read a range of SNPs with verbose output
dosages <- getvcfsnp(vcf_summary, snp_indices = 1:100, verbose = TRUE)
}
}
