% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readvcf.R
\name{summarizesnpinfo}
\alias{summarizesnpinfo}
\title{Create SNP Summary Data Frame}
\usage{
summarizesnpinfo(vcf_snps, verbose = TRUE)
}
\arguments{
\item{vcf_snps}{VCF object from getsnpinfo() function}

\item{verbose}{Logical; if TRUE, display informational messages. Default is TRUE.}
}
\value{
A list containing snp_df (data frame with SNP ID, chromosome, position, reference allele, and alternate allele) and rr (rowRanges object)
}
\description{
Create SNP Summary Data Frame
}
\examples{
\dontrun{
vcf_file <- "data/axiom_acs_aus_nf_chr21.vcf.gz"
snps <- getsnpinfo(vcf_file)
result <- summarizesnpinfo(snps)
snp_df <- result$snp_df
rr <- result$rr
}
}
