% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readvcf.R
\name{summarizevcf}
\alias{summarizevcf}
\title{Summarize VCF File}
\usage{
summarizevcf(vcf_file, genome = "hg19", verbose = TRUE)
}
\arguments{
\item{vcf_file}{Character string with path to VCF file (must be bgzipped)}

\item{genome}{Character string specifying genome build (e.g., "hg19", "hg38")}

\item{verbose}{Logical; if TRUE, display informational messages. Default is TRUE.}
}
\value{
A list containing:
\itemize{
\item vcf_file: The VCF file name
\item genome: The genome build used
\item snp_summary: The list returned from summarizesnpinfo (contains snp_df and rr)
\item sample_ids: Vector of sample IDs from the VCF file
}
}
\description{
Summarize VCF File
}
\examples{
\dontrun{
vcf_file <- "data/axiom_acs_aus_nf_chr21.vcf.gz"
summary <- summarizevcf(vcf_file)
summary$vcf_file
summary$genome
summary$snp_summary$snp_df
summary$sample_ids
}
}
